requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
confusable-homoglyphs==3.3.1
cryptography==45.0.7
Django==5.2.6
django-js-asset==3.1.2
django-mptt==0.18.0
django-registration==5.2.1
djangorestframework==3.16.1
idna==3.10
josepy==2.1.0
mozilla-django-oidc==4.0.1
psycopg2-binary==2.9.10
pycparser==2.23
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
